requirements.txt
flask
flask-sqlalchemy
flask-migrate
python-dotenv
flask-jwt-extended
werkzeug
pymysql
cryptography
google-genai